renamed SC client
